Increase sha1 slice size and adjust refine_uid:
refine_uid now returns a shorter string without uppercase letters.
If the uid is part of the URL, there is no case sensititvityi, so only one
case is needed. generate_uid returns more digits of the sha1 digest.
Within an order of magnitude, there is the same number of possibilites for
the return values of generate_uid and refine_uid